leftpane conversation  header revamp done.
