+++ ts/components/conversation/ContactName.tsx
@@ -15,17 +15,17 @@ type Props = {
 };
 
 export const ContactName = (props: Props) => {
-  const { pubkey, name, profileName, module, boldProfileName, compact, shouldShowPubkey } = props;
+  const { pubkey, name, profileName, module, compact, shouldShowPubkey } = props;
   const prefix = module ? module : 'module-contact-name';
 
   const convoName = useConversationUsernameOrShorten(pubkey);
   const isPrivate = useIsPrivate(pubkey);
   const shouldShowProfile = Boolean(convoName || profileName || name);
-  const styles = (boldProfileName
-    ? {
-        fontWeight: 'bold',
-      }
-    : {}) as React.CSSProperties;
+  // const styles = (boldProfileName
+  //   ? {
+  //       fontWeight: 'bold',
+  //     }
+  //   : {}) as React.CSSProperties;
   const textProfile = profileName || name || convoName || window.i18n('anonymous');
 
   return (
@@ -35,7 +35,7 @@ export const ContactName = (props: Props) => {
       data-testid={`${prefix}__profile-name`}
     >
       {shouldShowProfile ? (
-        <span style={styles as any} className={`${prefix}__profile-name`}>
+        <span  className={`${prefix}__profile-name`}>
           <Emojify text={textProfile} sizeClass="small" isGroup={!isPrivate} />
         </span>
       ) : null}
