+++ ts/components/icon/Icons.tsx
@@ -211,8 +211,8 @@ export const icons = {
   },
   bell: {
     path:
-      'M2.117 0a.396.396 0 00-.397.397v.18C.963.757.53 1.434.53 2.25v1.323l-.53.53v.264h4.233V4.1l-.529-.53v-.223h-.29c-.066 0-.132-.006-.197-.015v.546l-2.159-.042V2.249c0-.656.4-1.19 1.059-1.19l.064.003c.119-.181.278-.334.463-.448a1.608 1.608 0 00-.13-.036v-.18A.396.396 0 002.117 0zm-.53 4.63a.53.53 0 001.058 0z M3.355.578a1.267 1.267 0 000 2.534h.634v-.254h-.634c-.55 0-1.013-.464-1.013-1.013 0-.55.463-1.014 1.013-1.014.55 0 1.014.464 1.014 1.014v.18c0 .1-.09.2-.19.2s-.19-.1-.19-.2v-.18a.634.634 0 10-.185.447.47.47 0 00.375.186c.25 0 .443-.203.443-.452v-.181c0-.7-.567-1.267-1.267-1.267zm0 1.647a.38.38 0 110-.76.38.38 0 010 .76z',
-    viewBox: '0 0 4.622 5.159',
+    "M16 8.35996V10.544C16 11.9328 14.8807 13.0587 13.5 13.0587H9.68787L6.62533 15.7541C5.99168 16.3119 5 15.8593 5 15.0123V13.0587H4.5C3.11929 13.0587 2 11.9328 2 10.544V5.51468C2 4.12586 3.11929 3 4.5 3H8.75777C8.55127 3.31032 8.3819 3.64775 8.25606 4.00587H4.5C3.67157 4.00587 3 4.68138 3 5.51468V10.544C3 11.3773 3.67157 12.0528 4.5 12.0528H6V14.968L9.31213 12.0528H13.5C14.3284 12.0528 15 11.3773 15 10.544V9.27888C15.3776 9.02461 15.7149 8.71437 16 8.35996ZM9.99958 6.5C9.99958 5.11905 11.1191 3.99957 12.5 3.99957C13.881 3.99957 15.0004 5.11905 15.0004 6.5V6.50233C14.9986 6.7774 14.7751 6.99978 14.4999 6.99978C14.2235 6.99978 13.9993 6.77568 13.9993 6.49923V5.49909C13.9993 5.22305 13.7755 4.9993 13.4996 4.9993C13.3612 4.9993 13.236 5.05551 13.1455 5.14633C12.95 5.05293 12.7311 5.00064 12.5 5.00064C11.6719 5.00064 11.0006 5.67192 11.0006 6.5C11.0006 7.32807 11.6719 7.99936 12.5 7.99936C12.8841 7.99936 13.2345 7.85491 13.4998 7.61736C13.7652 7.85492 14.1157 7.99936 14.4999 7.99936C15.3286 7.99936 16 7.32732 16 6.49899C15.9995 4.56645 14.4327 3 12.5 3C10.567 3 9 4.567 9 6.5C9 8.43299 10.567 10 12.5 10C12.918 10 13.301 9.9255 13.6691 9.79275C13.9288 9.6991 14.0633 9.41269 13.9697 9.15303C13.8761 8.89338 13.5896 8.75881 13.33 8.85246C13.0622 8.94902 12.7938 9.00043 12.5 9.00043C11.1191 9.00043 9.99958 7.88095 9.99958 6.5ZM12.5 6.99978C12.224 6.99978 12.0002 6.77602 12.0002 6.5C12.0002 6.22398 12.224 6.00021 12.5 6.00021C12.776 6.00021 12.9998 6.22398 12.9998 6.5C12.9998 6.77602 12.776 6.99978 12.5 6.99978Z",
+    viewBox: '0 0 18 18',
     ratio: 1,
   },
   brand: {
@@ -545,8 +545,8 @@ export const icons = {
   },
   mute: {
     path:
-      'M12,22c1.1,0 2,-0.9 2,-2h-4c0,1.1 0.9,2 2,2zM12,6.5c2.49,0 4,2.02 4,4.5v0.1l2,2L18,11c0,-3.07 -1.63,-5.64 -4.5,-6.32L13.5,4c0,-0.83 -0.67,-1.5 -1.5,-1.5s-1.5,0.67 -1.5,1.5v0.68c-0.24,0.06 -0.47,0.15 -0.69,0.23l1.64,1.64c0.18,-0.02 0.36,-0.05 0.55,-0.05zM5.41,3.35L4,4.76l2.81,2.81C6.29,8.57 6,9.74 6,11v5l-2,2v1h14.24l1.74,1.74 1.41,-1.41L5.41,3.35zM16,17L8,17v-6c0,-0.68 0.12,-1.32 0.34,-1.9L16,16.76L16,17z',
-    viewBox: '0 0 24 24',
+    "M9.16992 2.59675C9.8075 2.03045 10.8125 2.48346 10.8125 3.33582V14.6639C10.8125 15.5152 9.80566 15.9707 9.16915 15.4023C9.16902 15.4022 9.16889 15.4021 9.16876 15.402L5.69272 12.3124H3.1875C2.29026 12.3124 1.5625 11.5846 1.5625 10.6874V7.31238C1.5625 6.41515 2.29026 5.68738 3.1875 5.68738H5.69272L9.16978 2.59687L9.16992 2.59675ZM9.8125 3.36352L6.21498 6.5611C6.12346 6.64245 6.00526 6.68738 5.88281 6.68738H3.1875C2.84255 6.68738 2.5625 6.96743 2.5625 7.31238V10.6874C2.5625 11.0323 2.84255 11.3124 3.1875 11.3124H5.88281C6.00526 11.3124 6.12346 11.3573 6.21498 11.4387L9.8125 14.6362V3.36352ZM12.5839 6.95883C12.7792 6.76357 13.0958 6.76357 13.2911 6.95883L14.625 8.29278L15.9589 6.95883C16.1542 6.76357 16.4708 6.76357 16.6661 6.95883C16.8613 7.15409 16.8613 7.47067 16.6661 7.66594L15.3321 8.99988L16.6661 10.3338C16.8613 10.5291 16.8613 10.8457 16.6661 11.0409C16.4708 11.2362 16.1542 11.2362 15.9589 11.0409L14.625 9.70699L13.2911 11.0409C13.0958 11.2362 12.7792 11.2362 12.5839 11.0409C12.3887 10.8457 12.3887 10.5291 12.5839 10.3338L13.9179 8.99988L12.5839 7.66594C12.3887 7.47067 12.3887 7.15409 12.5839 6.95883Z",
+    viewBox: '0 0 18 18',
     ratio: 1,
   },
   notification: {
