+++ ts/components/leftpane/conversation-list-item/HeaderItem.tsx
@@ -1,20 +1,18 @@
 import classNames from 'classnames';
-import React, { 
-  // useCallback,
-   useContext } from 'react';
+import React, { useCallback, useContext } from 'react';
 import { useSelector } from 'react-redux';
 import styled from 'styled-components';
-// import { getFirstUnreadMessageWithMention } from '../../../data/data';
+import { getFirstUnreadMessageWithMention } from '../../../data/data';
 import { useConversationPropsById, useIsPinned } from '../../../hooks/useParamSelector';
-// import { UserUtils } from '../../../bchat/utils';
-// import {
-//   openConversationToSpecificMessage,
-//   openConversationWithMessages,
-// } from '../../../state/ducks/conversations';
+import { UserUtils } from '../../../bchat/utils';
+import {
+  openConversationToSpecificMessage,
+  openConversationWithMessages,
+} from '../../../state/ducks/conversations';
 import { SectionType } from '../../../state/ducks/section';
 import { isSearching } from '../../../state/selectors/search';
 import { getFocusedSection } from '../../../state/selectors/section';
-import { Timestamp } from '../../conversation/Timestamp';
+// import { Timestamp } from '../../conversation/Timestamp';
 import { BchatIcon } from '../../icon';
 import { ContextConversationId } from './ConversationListItem';
 import { UserItem } from './UserItem';
@@ -31,13 +29,9 @@ const NotificationSettingIcon = (props: { isMessagesSection: boolean }) => {
     case 'all':
       return null;
     case 'disabled':
-      return (
-        <BchatIcon iconType="mute" iconColor={'var(--color-text-subtle)'} iconSize="small" />
-      );
+      return <BchatIcon iconType="mute" iconColor={'#A7A7BA'} iconSize="medium" />;
     case 'mentions_only':
-      return (
-        <BchatIcon iconType="bell" iconColor={'var(--color-text-subtle)'} iconSize="small" />
-      );
+      return <BchatIcon iconType="bell" iconColor={'var(--color-text-subtle)'} iconSize="medium" />;
     default:
       return null;
   }
@@ -51,11 +45,19 @@ const StyledConversationListItemIconWrapper = styled.div`
   display: flex;
   flex-direction: row;
 `;
+const PinWrapper = styled.div`
+  position: absolute;
+  top: -6px;
+  right: 12px;
+  padding:1px 2px;
+  background-color: #2f8fff;
+  border-radius: 12px;
+`;
 
 function useHeaderItemProps(conversationId: string) {
   const convoProps = useConversationPropsById(conversationId);
   // console.log();
-  
+
   if (!convoProps) {
     return null;
   }
@@ -74,7 +76,9 @@ const ListItemIcons = () => {
 
   const pinIcon =
     isMessagesSection && isPinned ? (
-      <BchatIcon iconType="pin" iconColor={'var(--color-text-subtle)'} iconSize="small" />
+      <PinWrapper>
+        <BchatIcon iconType="pin" iconColor={'#F0F0F0'} iconSize={14} />
+      </PinWrapper>
     ) : null;
   return (
     <StyledConversationListItemIconWrapper>
@@ -84,34 +88,34 @@ const ListItemIcons = () => {
   );
 };
 
-// const MentionAtSymbol = styled.span`
-//   background-color: var(--color-accent);
+const MentionAtSymbol = styled.span`
+  background-color: var(--color-accent);
 
-//   color: black;
-//   text-align: center;
-//   margin-top: 0px;
-//   margin-bottom: 0px;
-//   padding-top: 1px;
-//   padding-inline-start: 3px;
-//   padding-inline-end: 3px;
+  color: black;
+  text-align: center;
+  margin-top: 0px;
+  margin-bottom: 0px;
+  padding-top: 1px;
+  padding-inline-start: 3px;
+  padding-inline-end: 3px;
 
-//   position: static;
-//   margin-inline-start: 5px;
+  position: static;
+  margin-inline-start: 5px;
 
-//   font-weight: 300;
-//   font-size: 11px;
-//   letter-spacing: 0.25px;
+  font-weight: 300;
+  font-size: 11px;
+  letter-spacing: 0.25px;
 
-//   height: 16px;
-//   min-width: 16px;
-//   border-radius: 8px;
-//   /* transition: filter 0.25s linear; */
-//   cursor: pointer;
+  height: 16px;
+  min-width: 16px;
+  border-radius: 8px;
+  /* transition: filter 0.25s linear; */
+  cursor: pointer;
 
-//   :hover {
-//     filter: grayscale(0.7);
-//   }
-// `;
+  :hover {
+    filter: grayscale(0.7);
+  }
+`;
 
 export const ConversationListItemHeaderItem = () => {
   const conversationId = useContext(ContextConversationId);
@@ -120,54 +124,59 @@ export const ConversationListItemHeaderItem = () => {
 
   const convoProps = useHeaderItemProps(conversationId);
 
-  // const openConvoToLastMention = useCallback(
-  //   async (e: React.MouseEvent<HTMLDivElement>) => {
-  //     e.stopPropagation();
-  //     e.preventDefault();
+  const openConvoToLastMention = useCallback(
+    async (e: React.MouseEvent<HTMLDivElement>) => {
+      e.stopPropagation();
+      e.preventDefault();
 
-  //     // mousedown is invoked sooner than onClick, but for both right and left click
-  //     if (e.button === 0) {
-  //       const oldestMessageUnreadWithMention =
-  //         (await getFirstUnreadMessageWithMention(
-  //           conversationId,
-  //           UserUtils.getOurPubKeyStrFromCache()
-  //         )) || null;
-  //       if (oldestMessageUnreadWithMention) {
-  //         await openConversationToSpecificMessage({
-  //           conversationKey: conversationId,
-  //           messageIdToNavigateTo: oldestMessageUnreadWithMention,
-  //           shouldHighlightMessage: true,
-  //         });
-  //       } else {
-  //         window.log.info('cannot open to latest mention as no unread mention are found');
-  //         await openConversationWithMessages({
-  //           conversationKey: conversationId,
-  //           messageId: null,
-  //         });
-  //       }
-  //     }
-  //   },
-  //   [conversationId]
-  // );
+      // mousedown is invoked sooner than onClick, but for both right and left click
+      if (e.button === 0) {
+        const oldestMessageUnreadWithMention =
+          (await getFirstUnreadMessageWithMention(
+            conversationId,
+            UserUtils.getOurPubKeyStrFromCache()
+          )) || null;
+        if (oldestMessageUnreadWithMention) {
+          await openConversationToSpecificMessage({
+            conversationKey: conversationId,
+            messageIdToNavigateTo: oldestMessageUnreadWithMention,
+            shouldHighlightMessage: true,
+          });
+        } else {
+          window.log.info('cannot open to latest mention as no unread mention are found');
+          await openConversationWithMessages({
+            conversationKey: conversationId,
+            messageId: null,
+          });
+        }
+      }
+    },
+    [conversationId]
+  );
 
   if (!convoProps) {
     return null;
   }
-  const { unreadCount, 
-    //  mentionedUs,
-     activeAt
-     } = convoProps;
-
-  // let atSymbol = null;
-  // let unreadCountDiv = null;
-  // if (unreadCount > 0) {
-  //   atSymbol = mentionedUs ? (
-  //     <MentionAtSymbol title="Open to latest mention" onMouseDown={openConvoToLastMention}>
-  //       @
-  //     </MentionAtSymbol>
-  //   ) : null;
-  //   unreadCountDiv = <p className="module-conversation-list-item__unread-count">{unreadCount>99?"99+":unreadCount}</p>;
-  // }
+  const {
+    unreadCount,
+    mentionedUs,
+    //  activeAt
+  } = convoProps;
+  // const unreadCount = 15;
+  let atSymbol = null;
+  let unreadCountDiv = null;
+  if (unreadCount > 0) {
+    atSymbol = mentionedUs ? (
+      <MentionAtSymbol title="Open to latest mention" onMouseDown={openConvoToLastMention}>
+        @
+      </MentionAtSymbol>
+    ) : null;
+    unreadCountDiv = (
+      <p className="module-conversation-list-item__unread-count">
+        {unreadCount > 99 ? '99+' : unreadCount}
+      </p>
+    );
+  }
 
   return (
     <div className="module-conversation-list-item__header">
@@ -178,14 +187,14 @@ export const ConversationListItemHeaderItem = () => {
         )}
       >
         <UserItem />
-      
-
       </div>
-      <ListItemIcons />
-{/* 
-      {unreadCountDiv}
-      {atSymbol} */}
-      <Timestamp timestamp={activeAt} isConversationListItem={true} momentFromNow={true} />
+      <div className="module-conversation-list-item_unread-count_wrapper">
+        <ListItemIcons />
+        {atSymbol}
+        {unreadCountDiv}
+      </div>
+
+      {/* <Timestamp timestamp={activeAt} isConversationListItem={true} momentFromNow={true} /> */}
       {!isSearchingMode && (
         <div
           className={classNames(
